Move user auth to middleware
